+++ packages/optionmenu/OptionDivider.svelte
@@ -1,0 +1,1 @@
+<hr class="option-menu__list-item__divider" />