+++ packages/optionmenu/base.js
@@ -1,3 +1,5 @@
 import OptionMenu from "./OptionMenu.svelte";
+import Option from "./Option.svelte";
+import OptionDivider from "./OptionDivider.svelte";
 
-export { OptionMenu };
+export { OptionMenu, Option, OptionDivider };